Move matchUtils to their own directory
In order to break and avoid dependency cycles we need to move match
utils to its own folder. The file that was previously 'matchUtils.ts'
will now be the index of the directory. This will allow us to store
single-util files in the dir in future.
