+++ server/controllers/match/search/spaceSearchController.ts
@@ -6,7 +6,7 @@ import matchPaths from '../../../paths/match'
 import { PlacementRequestService } from '../../../services'
 import SpaceService from '../../../services/spaceService'
 
-import { startDateObjFromParams } from '../../../utils/matchUtils'
+import { startDateObjFromParams } from '../../../utils/match'
 import { objectIfNotEmpty } from '../../../utils/utils'
 
 export default class BedSearchController {
