+++ server/utils/nunjucksSetup.ts
@@ -51,7 +51,7 @@ import * as FormUtils from './formUtils'
 import * as UserUtils from './users'
 import * as TaskUtils from './tasks'
 import * as PlacementRequestUtils from './placementRequests'
-import * as MatchUtils from './matchUtils'
+import * as MatchUtils from './match'
 import * as SummaryListUtils from './applications/summaryListUtils'
 import * as BedUtils from './bedUtils'
 import * as LostBedUtils from './lostBedUtils'
