+++ server/utils/placementRequests/adminSummary.ts
@@ -4,7 +4,7 @@ import { SummaryList, SummaryListItem } from '../../@types/ui'
 import { allReleaseTypes } from '../applications/releaseTypeUtils'
 import { withdrawnStatusTag } from '../applications/utils'
 import { DateFormats } from '../dateUtils'
-import { placementDates, placementLength } from '../matchUtils'
+import { placementDates, placementLength } from '../match'
 import paths from '../../paths/apply'
 
 export const adminSummary = (placementRequest: PlacementRequestDetail): SummaryList => {
